Refactor unit tests folder structure.
