+++ tests/test_models/test_backbones/test_shufflenet_v1.py
