+++ tests/test_models/test_backbones/test_shufflenet_v2.py
