+++ tests/test_models/test_backbones/test_swin_transformer.py
